+++ Assets/Scripts/GameCode/FootstepSensor.cs
@@ -19,30 +19,30 @@ class FootstepSensor {
 	 * Returns -1 if there are no guards close. Returns 0 if guard is at PlayerAgent.
 	 * Otherwise returns value up to GameConstants.FootstepSensorBehindPixels (distance of closest GuardAgent)
 	 */
-	public short getSoundBehindPlayer() {
-		short pLocX = GameEngine.player.LocationX;
-		short pLocY = GameEngine.player.LocationY;
+	public int getSoundBehindPlayer() {
+		int pLocX = GameEngine.player.LocationX;
+		int pLocY = GameEngine.player.LocationY;
 
-		short cloestGuard = (short)(GameConstants.FootstepSensorBehindPixels + 10);
-		short distance = cloestGuard;
+		double cloestGuard = (GameConstants.FootstepSensorBehindPixels + 10);
+		double distance = cloestGuard;
 		for (int i = 0; i<GameEngine.guards.Count; i++) {
 			if (GameEngine.player.MovingDirection == GameConstants.Up && GameEngine.player.LocationY > GameEngine.guards[i].LocationY) { // PlayerAgent facing up
-				distance = (short)(Math.Pow(pLocX-GameEngine.guards[i].LocationX,2)+Math.Pow(pLocY-GameEngine.guards[i].LocationY,2));
+				distance = (Math.Pow(pLocX-GameEngine.guards[i].LocationX,2)+Math.Pow(pLocY-GameEngine.guards[i].LocationY,2));
 				if (distance < cloestGuard) {
 					cloestGuard = distance;
 				}
 			} else if (GameEngine.player.MovingDirection == GameConstants.Down && GameEngine.player.LocationY < GameEngine.guards[i].LocationY) { // PlayerAgent facing down
-				distance = (short)(Math.Pow(pLocX-GameEngine.guards[i].LocationX,2)+Math.Pow(pLocY-GameEngine.guards[i].LocationY,2));
+				distance = (Math.Pow(pLocX-GameEngine.guards[i].LocationX,2)+Math.Pow(pLocY-GameEngine.guards[i].LocationY,2));
 				if (distance < cloestGuard) {
 					cloestGuard = distance;
 				}
 			} else if (GameEngine.player.MovingDirection == GameConstants.Left && GameEngine.player.LocationX < GameEngine.guards[i].LocationX) { // PlayerAgent facing left
-				distance =(short) (Math.Pow(pLocX-GameEngine.guards[i].LocationX,2)+Math.Pow(pLocY-GameEngine.guards[i].LocationY,2));
+				distance = (Math.Pow(pLocX-GameEngine.guards[i].LocationX,2)+Math.Pow(pLocY-GameEngine.guards[i].LocationY,2));
 				if (distance < cloestGuard) {
 					cloestGuard = distance;
 				}
 			} else if (GameEngine.player.MovingDirection == GameConstants.Right && GameEngine.player.LocationX > GameEngine.guards[i].LocationX) { // PlayerAgent facing right
-				distance = (short)(Math.Pow(pLocX-GameEngine.guards[i].LocationX,2)+Math.Pow(pLocY-GameEngine.guards[i].LocationY,2));
+				distance = (Math.Pow(pLocX-GameEngine.guards[i].LocationX,2)+Math.Pow(pLocY-GameEngine.guards[i].LocationY,2));
 				if (distance < cloestGuard) {
 					cloestGuard = distance;
 				}
@@ -53,7 +53,7 @@ class FootstepSensor {
 			return -1;
 		}
 
-		return distance;
+		return (int)distance;
 	}
 
 	/* 
@@ -62,30 +62,30 @@ class FootstepSensor {
 	 * Returns -1 if there are no guards close. Returns 0 if guard is at PlayerAgent.
 	 * Otherwise returns value up to GameConstants.FootstepSensorFrontPixels (distance of closest GuardAgent)
 	 */
-    public short getSoundInFrontOfPlayer() {
-		short pLocX = GameEngine.player.LocationX;
-		short pLocY = GameEngine.player.LocationY;
+	public int getSoundInFrontOfPlayer() {
+		int pLocX = GameEngine.player.LocationX;
+		int pLocY = GameEngine.player.LocationY;
 		
-		short cloestGuard = (short)(GameConstants.FootstepSensorFrontPixels + 10);
-		short distance = cloestGuard;
+		double cloestGuard = (GameConstants.FootstepSensorFrontPixels + 10);
+		double distance = cloestGuard;
 		for (int i = 0; i<GameEngine.guards.Count; i++) {
 			if (GameEngine.player.MovingDirection == GameConstants.Up && GameEngine.player.LocationY < GameEngine.guards[i].LocationY) { // PlayerAgent facing up
-				distance = (short)(Math.Pow(pLocX-GameEngine.guards[i].LocationX,2)+Math.Pow(pLocY-GameEngine.guards[i].LocationY,2));
+				distance = (Math.Pow(pLocX-GameEngine.guards[i].LocationX,2)+Math.Pow(pLocY-GameEngine.guards[i].LocationY,2));
 				if (distance < cloestGuard) {
 					cloestGuard = distance;
 				}
 			} else if (GameEngine.player.MovingDirection == GameConstants.Down && GameEngine.player.LocationY > GameEngine.guards[i].LocationY) { // PlayerAgent facing down
-				distance = (short)(Math.Pow(pLocX-GameEngine.guards[i].LocationX,2)+Math.Pow(pLocY-GameEngine.guards[i].LocationY,2));
+				distance = (Math.Pow(pLocX-GameEngine.guards[i].LocationX,2)+Math.Pow(pLocY-GameEngine.guards[i].LocationY,2));
 				if (distance < cloestGuard) {
 					cloestGuard = distance;
 				}
 			} else if (GameEngine.player.MovingDirection == GameConstants.Left && GameEngine.player.LocationX > GameEngine.guards[i].LocationX) { // PlayerAgent facing left
-				distance = (short)(Math.Pow(pLocX-GameEngine.guards[i].LocationX,2)+Math.Pow(pLocY-GameEngine.guards[i].LocationY,2));
+				distance = (Math.Pow(pLocX-GameEngine.guards[i].LocationX,2)+Math.Pow(pLocY-GameEngine.guards[i].LocationY,2));
 				if (distance < cloestGuard) {
 					cloestGuard = distance;
 				}
 			} else if (GameEngine.player.MovingDirection == GameConstants.Right && GameEngine.player.LocationX < GameEngine.guards[i].LocationX) { // PlayerAgent facing right
-				distance = (short)(Math.Pow(pLocX-GameEngine.guards[i].LocationX,2)+Math.Pow(pLocY-GameEngine.guards[i].LocationY,2));
+				distance = (Math.Pow(pLocX-GameEngine.guards[i].LocationX,2)+Math.Pow(pLocY-GameEngine.guards[i].LocationY,2));
 				if (distance < cloestGuard) {
 					cloestGuard = distance;
 				}
@@ -96,7 +96,7 @@ class FootstepSensor {
 			return -1;
 		}
 		
-		return distance;
+		return (int)distance;
 	}
 }
